refactor: 타임라인 관련 컴포넌트 아토믹 패턴 재분류
