+++ features/main/time-line/atoms/index.ts
@@ -1,1 +1,1 @@
-export * from './time-line-card';
+export * from './swim-record-chart';
