+++ features/main/time-line/molecules/index.ts
@@ -1,1 +1,1 @@
-export * from './time-line';
+export * from './time-line-card';
