+++ features/main/time-line/organisms/index.ts
@@ -1,0 +1,1 @@
+export * from './time-line';
